data-pipeline/requirements.txt
requests>=2.31
pydantic>=2.11
typer>=0.9
python-slugify>=8.0
rich>=13
rtoml>=0.12
pandas>=2.0
